+++ courier/template/email/recovery_code_valid.go
@@ -18,9 +18,10 @@ type (
 		model *RecoveryCodeValidModel
 	}
 	RecoveryCodeValidModel struct {
-		To           string
-		RecoveryCode string
-		Identity     map[string]interface{}
+		To           string                 `json:"to"`
+		RecoveryCode string                 `json:"recovery_code"`
+		Identity     map[string]interface{} `json:"identity"`
+		RequestURL   string                 `json:"request_url"`
 	}
 )