+++ courier/template/email/recovery_invalid.go
@@ -18,7 +18,8 @@ type (
 		m *RecoveryInvalidModel
 	}
 	RecoveryInvalidModel struct {
-		To string
+		To         string `json:"to"`
+		RequestURL string `json:"request_url"`
 	}
 )