+++ courier/template/email/verification_code_valid.go
@@ -18,10 +18,11 @@ type (
 		m *VerificationCodeValidModel
 	}
 	VerificationCodeValidModel struct {
-		To               string
-		VerificationURL  string
-		VerificationCode string
-		Identity         map[string]interface{}
+		To               string                 `json:"to"`
+		VerificationURL  string                 `json:"verification_url"`
+		VerificationCode string                 `json:"verification_code"`
+		Identity         map[string]interface{} `json:"identity"`
+		RequestURL       string                 `json:"request_url"`
 	}
 )