+++ courier/template/email/verification_invalid.go
@@ -18,7 +18,8 @@ type (
 		m *VerificationInvalidModel
 	}
 	VerificationInvalidModel struct {
-		To string
+		To         string `json:"to"`
+		RequestURL string `json:"request_url"`
 	}
 )