+++ courier/template/email/verification_valid.go
@@ -18,9 +18,10 @@ type (
 		m *VerificationValidModel
 	}
 	VerificationValidModel struct {
-		To              string
-		VerificationURL string
-		Identity        map[string]interface{}
+		To              string                 `json:"to"`
+		VerificationURL string                 `json:"verification_url"`
+		Identity        map[string]interface{} `json:"identity"`
+		RequestURL      string                 `json:"request_url"`
 	}
 )