+++ courier/template/sms/login_code_valid.go
@@ -17,9 +17,10 @@ type (
 		model *LoginCodeValidModel
 	}
 	LoginCodeValidModel struct {
-		To        string
-		LoginCode string
-		Identity  map[string]interface{}
+		To         string                 `json:"to"`
+		LoginCode  string                 `json:"login_code"`
+		Identity   map[string]interface{} `json:"identity"`
+		RequestURL string                 `json:"request_url"`
 	}
 )