+++ persistence/sql/persister_continuity.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/gofrs/uuid"
 
+	"github.com/ory/x/otelx"
 	"github.com/ory/x/sqlcon"
 
 	"github.com/ory/kratos/continuity"
@@ -19,17 +20,17 @@ import (
 
 var _ continuity.Persister = new(Persister)
 
-func (p *Persister) SaveContinuitySession(ctx context.Context, c *continuity.Container) error {
+func (p *Persister) SaveContinuitySession(ctx context.Context, c *continuity.Container) (err error) {
 	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.SaveContinuitySession")
-	defer span.End()
+	defer otelx.End(span, &err)
 
 	c.NID = p.NetworkID(ctx)
 	return sqlcon.HandleError(p.GetConnection(ctx).Create(c))
 }
 
-func (p *Persister) GetContinuitySession(ctx context.Context, id uuid.UUID) (*continuity.Container, error) {
+func (p *Persister) GetContinuitySession(ctx context.Context, id uuid.UUID) (_ *continuity.Container, err error) {
 	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.GetContinuitySession")
-	defer span.End()
+	defer otelx.End(span, &err)
 
 	var c continuity.Container
 	if err := p.GetConnection(ctx).Where("id = ? AND nid = ?", id, p.NetworkID(ctx)).First(&c); err != nil {
@@ -38,9 +39,9 @@ func (p *Persister) GetContinuitySession(ctx context.Context, id uuid.UUID) (*co
 	return &c, nil
 }
 
-func (p *Persister) DeleteContinuitySession(ctx context.Context, id uuid.UUID) error {
+func (p *Persister) DeleteContinuitySession(ctx context.Context, id uuid.UUID) (err error) {
 	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.DeleteContinuitySession")
-	defer span.End()
+	defer otelx.End(span, &err)
 
 	if count, err := p.GetConnection(ctx).RawQuery(
 		//#nosec G201 -- TableName is static
@@ -53,9 +54,11 @@ func (p *Persister) DeleteContinuitySession(ctx context.Context, id uuid.UUID) e
 	return nil
 }
 
-func (p *Persister) DeleteExpiredContinuitySessions(ctx context.Context, expiresAt time.Time, limit int) error {
+func (p *Persister) DeleteExpiredContinuitySessions(ctx context.Context, expiresAt time.Time, limit int) (err error) {
+	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.DeleteExpiredContinuitySessions")
+	defer otelx.End(span, &err)
 	//#nosec G201 -- TableName is static
-	err := p.GetConnection(ctx).RawQuery(fmt.Sprintf(
+	err = p.GetConnection(ctx).RawQuery(fmt.Sprintf(
 		"DELETE FROM %s WHERE id in (SELECT id FROM (SELECT id FROM %s c WHERE expires_at <= ? and nid = ? ORDER BY expires_at ASC LIMIT %d ) AS s )",
 		new(continuity.Container).TableName(ctx),
 		new(continuity.Container).TableName(ctx),