+++ persistence/sql/persister_verification.go
@@ -16,6 +16,7 @@ import (
 	"github.com/gobuffalo/pop/v6"
 	"github.com/gofrs/uuid"
 
+	"github.com/ory/x/otelx"
 	"github.com/ory/x/sqlcon"
 
 	"github.com/ory/kratos/selfservice/flow/verification"
@@ -24,9 +25,9 @@ import (
 
 var _ verification.FlowPersister = new(Persister)
 
-func (p *Persister) CreateVerificationFlow(ctx context.Context, r *verification.Flow) error {
+func (p *Persister) CreateVerificationFlow(ctx context.Context, r *verification.Flow) (err error) {
 	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.CreateVerificationFlow")
-	defer span.End()
+	defer otelx.End(span, &err)
 
 	r.NID = p.NetworkID(ctx)
 	// This should not create the request eagerly because otherwise we might accidentally create an address
@@ -34,9 +35,9 @@ func (p *Persister) CreateVerificationFlow(ctx context.Context, r *verification.
 	return p.GetConnection(ctx).Create(r)
 }
 
-func (p *Persister) GetVerificationFlow(ctx context.Context, id uuid.UUID) (*verification.Flow, error) {
+func (p *Persister) GetVerificationFlow(ctx context.Context, id uuid.UUID) (_ *verification.Flow, err error) {
 	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.GetVerificationFlow")
-	defer span.End()
+	defer otelx.End(span, &err)
 
 	var r verification.Flow
 	if err := p.GetConnection(ctx).Where("id = ? AND nid = ?", id, p.NetworkID(ctx)).First(&r); err != nil {
@@ -46,18 +47,18 @@ func (p *Persister) GetVerificationFlow(ctx context.Context, id uuid.UUID) (*ver
 	return &r, nil
 }
 
-func (p *Persister) UpdateVerificationFlow(ctx context.Context, r *verification.Flow) error {
+func (p *Persister) UpdateVerificationFlow(ctx context.Context, r *verification.Flow) (err error) {
 	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.UpdateVerificationFlow")
-	defer span.End()
+	defer otelx.End(span, &err)
 
 	cp := *r
 	cp.NID = p.NetworkID(ctx)
 	return update.Generic(ctx, p.GetConnection(ctx), p.r.Tracer(ctx).Tracer(), cp)
 }
 
-func (p *Persister) CreateVerificationToken(ctx context.Context, token *link.VerificationToken) error {
+func (p *Persister) CreateVerificationToken(ctx context.Context, token *link.VerificationToken) (err error) {
 	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.CreateVerificationToken")
-	defer span.End()
+	defer otelx.End(span, &err)
 
 	t := token.Token
 	token.Token = p.hmacValue(ctx, t)
@@ -72,9 +73,9 @@ func (p *Persister) CreateVerificationToken(ctx context.Context, token *link.Ver
 	return nil
 }
 
-func (p *Persister) UseVerificationToken(ctx context.Context, fID uuid.UUID, token string) (*link.VerificationToken, error) {
+func (p *Persister) UseVerificationToken(ctx context.Context, fID uuid.UUID, token string) (_ *link.VerificationToken, err error) {
 	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.UseVerificationToken")
-	defer span.End()
+	defer otelx.End(span, &err)
 
 	var rt link.VerificationToken
 
@@ -109,18 +110,20 @@ func (p *Persister) UseVerificationToken(ctx context.Context, fID uuid.UUID, tok
 	return &rt, nil
 }
 
-func (p *Persister) DeleteVerificationToken(ctx context.Context, token string) error {
+func (p *Persister) DeleteVerificationToken(ctx context.Context, token string) (err error) {
 	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.DeleteVerificationToken")
-	defer span.End()
+	defer otelx.End(span, &err)
 
 	nid := p.NetworkID(ctx)
 	//#nosec G201 -- TableName is static
 	return p.GetConnection(ctx).RawQuery(fmt.Sprintf("DELETE FROM %s WHERE token=? AND nid = ?", new(link.VerificationToken).TableName(ctx)), token, nid).Exec()
 }
 
-func (p *Persister) DeleteExpiredVerificationFlows(ctx context.Context, expiresAt time.Time, limit int) error {
+func (p *Persister) DeleteExpiredVerificationFlows(ctx context.Context, expiresAt time.Time, limit int) (err error) {
+	ctx, span := p.r.Tracer(ctx).Tracer().Start(ctx, "persistence.sql.DeleteExpiredVerificationFlows")
+	defer otelx.End(span, &err)
 	//#nosec G201 -- TableName is static
-	err := p.GetConnection(ctx).RawQuery(fmt.Sprintf(
+	err = p.GetConnection(ctx).RawQuery(fmt.Sprintf(
 		"DELETE FROM %s WHERE id in (SELECT id FROM (SELECT id FROM %s c WHERE expires_at <= ? and nid = ? ORDER BY expires_at ASC LIMIT %d ) AS s )",
 		new(verification.Flow).TableName(ctx),
 		new(verification.Flow).TableName(ctx),